Simple control over serialization...BUT
It turns out that for two reasons it is a bad idea to have a 'nolen'
version of decode.  The first is related to the header processing.
Previously, we relied on knowledge of BSON to determine the length
of the header (by previewing the first four bytes).  This is not good.
Second, it turns out that the compactness of msgpack comes partly from
the fact that it doesn't store lengths and it throws an exception for
extra data.  Not having lengths is also an issue when dealing with
compression (where it is required).

For this reason, the format needs to be refactored such that the lengths
of all data to be deserialized must be maintained outside the serialized
data.  I added this as a separate issue (#12).