Lots of excellent improvements
Added support for type constraints on columns in tables (and tests to test it).  This
closes #2.

Added tests for non-float types (int, bool and str).  This required me to deal with
places where flaots were assumed (e.g. scaling of alias variables).  The tests now
exercise a wide enough range of conditions that I'm convinced the files can be
quite heterogenous.