Improved fonts
These fonts work much better for 1024x768 screens.  However, the problem is that they
are an issue for 16:9 screens.  Since the presentation will be given on a 4:3 (1024x768)
screen, I've optimized them for that.  But ideally, flowtime.js should have a better
scaling algorithm so that it doesn't matter as much.