Lengths are no longer handled specially
Lengths are now implemented using msgpack.  This means that there is a potential savings
in bytes for things that are short AND we can use full 64-bit longs for lengths when
necessary.

This also completes the conversion to using msgpack for all data stored in these
files (issue #47 was already handled).